python_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dateutil==2.9.0
pytz==2024.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic-settings==2.6.1
websockets==14.1
openai==1.54.0
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
